requirements.txt
# Telegram
telethon>=1.36,<2.0

# Database
sqlalchemy>=2.0,<3.0

# HTTP
httpx>=0.27

# AI
openai>=1.50

# Image processing
Pillow>=10.0
imagehash>=4.3

# Scheduling
APScheduler>=3.10,<4.0

# Web dashboard
fastapi>=0.115
jinja2>=3.1
uvicorn>=0.32

# Config & utils
pyyaml>=6.0
loguru>=0.7
python-dotenv>=1.0

# AliExpress Affiliate API
python-aliexpress-api>=3.1

# Supabase (website database)
supabase>=2.0

# Testing
pytest>=8.0
pytest-asyncio>=0.24
respx>=0.22
